fix: bundle the CSS with the custom element
This prevents re-bundling problems when the CSS is referenced during the
downstream bundling process. vite will actually just leave a static
string when using `?url`, so the downstream bundler cannot reference the
CSS-File in the output anymore and will not re-bundle it, leaving the
string-URL as a dangling (and unusable) reference.

With this new approach Vite will instead just inline the CSS as a base64
string, which, while suboptimal, is perfectly servicable for what we
want to do.